chore: only create inter of not exist

+++ shell.nix
@@ -8,8 +8,10 @@ pkgs.mkShellNoCC {
 
   shellHook = ''
     mkdir -p ./app/fonts
-    cp -L "${
-      pkgs.google-fonts.override { fonts = [ "Inter" ]; }
-    }/share/fonts/truetype/Inter[opsz,wght].ttf" ./app/fonts/Inter.ttf
+    if [ ! -f ./app/fonts/Inter.ttf ]; then
+      cp -L "${
+        pkgs.google-fonts.override { fonts = [ "Inter" ]; }
+      }/share/fonts/truetype/Inter[opsz,wght].ttf" ./app/fonts/Inter.ttf
+    fi
   '';
 }
